+++ example/app/components/TaskManager.tsx
@@ -28,7 +28,7 @@ export const TaskManager: React.FC<{
       // of sync participants to successfully sync everything in the transaction, otherwise
       // no changes propagate and the transaction needs to start over when connectivity allows.
       realm.write(() => {
-        return new Task(realm, description, userId);
+        realm.create(Task, {description, userId});
       });
     },
     [realm, userId],